refactor code by routing handler
