--- apps/jan-api-gateway/application/app/domain/chat/constants.go
@@ -1,12 +1,0 @@
-package chat
-
-import "time"
-
-// Constants for streaming configuration
-const (
-	RequestTimeout    = 120 * time.Second
-	DataPrefix        = "data: "
-	DoneMarker        = "[DONE]"
-	ChannelBufferSize = 100
-	ErrorBufferSize   = 10
-)
