+++ apps/jan-api-gateway/application/cmd/server/wire_gen.go
@@ -8,9 +8,9 @@ package main
 
 import (
 	"context"
+
 	"menlo.ai/jan-api-gateway/app/domain/apikey"
 	"menlo.ai/jan-api-gateway/app/domain/auth"
-	"menlo.ai/jan-api-gateway/app/domain/chat"
 	"menlo.ai/jan-api-gateway/app/domain/conversation"
 	"menlo.ai/jan-api-gateway/app/domain/mcp/serpermcp"
 	"menlo.ai/jan-api-gateway/app/domain/organization"
@@ -28,22 +28,22 @@ import (
 	"menlo.ai/jan-api-gateway/app/infrastructure/database/repository/userrepo"
 	"menlo.ai/jan-api-gateway/app/infrastructure/inference"
 	"menlo.ai/jan-api-gateway/app/interfaces/http"
-	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1"
+	v1 "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1"
 	auth2 "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/auth"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/auth/google"
+	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/chat"
 	chat2 "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/chat"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/conversations"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/mcp"
-	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/mcp/mcp_impl"
+	mcpimpl "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/mcp/mcp_impl"
 	organization2 "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/organization"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/organization/projects"
-	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/organization/projects/api_keys"
+	apikeys "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/organization/projects/api_keys"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/responses"
-	"menlo.ai/jan-api-gateway/app/utils/httpclients/jan_inference"
-)
+	janinference "menlo.ai/jan-api-gateway/app/utils/httpclients/jan_inference"
 
-import (
 	_ "github.com/grafana/pyroscope-go/godeltaprof/http/pprof"
+
 	_ "net/http/pprof"
 )
 
@@ -74,9 +74,9 @@ func CreateApplication() (*Application, error) {
 	conversationRepository := conversationrepo.NewConversationGormRepository(transactionDatabase)
 	itemRepository := itemrepo.NewItemGormRepository(transactionDatabase)
 	conversationService := conversation.NewService(conversationRepository, itemRepository)
-	chatUseCase := chat.NewChatUseCase(inferenceProvider, conversationService)
-	streamingService := chat.NewStreamingService(inferenceProvider, conversationService)
-	completionAPI := chat2.NewCompletionAPI(apiKeyService, chatUseCase, streamingService, conversationService, authService)
+	completionNonStreamHandler := chat.NewCompletionNonStreamHandler(inferenceProvider, conversationService)
+	completionStreamHandler := chat.NewCompletionStreamHandler(inferenceProvider, conversationService)
+	completionAPI := chat2.NewCompletionAPI(completionNonStreamHandler, completionStreamHandler, conversationService, authService)
 	chatRoute := chat2.NewChatRoute(completionAPI, authService)
 	conversationAPI := conversations.NewConversationAPI(conversationService, authService)
 	modelAPI := v1.NewModelAPI()
